Move writeback types to `metabase-types/api`
