+++ frontend/src/metabase-types/api/index.ts
@@ -22,3 +22,4 @@ export * from "./slack";
 export * from "./table";
 export * from "./timeline";
 export * from "./user";
+export * from "./writeback";
