+++ frontend/src/metabase/writeback/components/ActionCreator/ActionCreatorHeader.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { t } from "ttag";
 
-import type { WritebackActionType } from "metabase/writeback/types";
+import type { WritebackActionType } from "metabase-types/api";
 
 import {
   Container,
