+++ frontend/src/metabase/writeback/selectors.ts
@@ -2,8 +2,8 @@ import Question from "metabase-lib/lib/Question";
 import { getSetting } from "metabase/selectors/settings";
 import { getMetadata } from "metabase/selectors/metadata";
 
+import type { WritebackQueryAction } from "metabase-types/api";
 import type { State } from "metabase-types/store";
-import type { WritebackQueryAction } from "./types";
 
 export function getWritebackEnabled(state: State) {
   return getSetting(state, "experimental-enable-actions");
